decision page architecture
