+++ flake-guard-app/src/client/components/Dashboard/UserDashboard.tsx
@@ -1,0 +1,9 @@
+const UserDashboard: React.FC = () => {
+  return (
+    <div>
+      <h1>USER DASHBOARD</h1>
+    </div>
+  );
+};
+
+export default UserDashboard;
