+++ flake-guard-app/src/server/routes/dasboardRouter.ts
@@ -1,0 +1,5 @@
+import express from 'express';
+
+const router = express.Router();
+
+export default router;
