+++ flake-guard-app/src/server/server.ts
@@ -4,8 +4,9 @@ import bodyParser from 'body-parser';
 import cors from 'cors';
 require('dotenv').config();
 const app = express();
-import npmRouter from './routes/npmRouter';
+import resultsRouter from './routes/resultsRouter';
 import tempDashRouter from './routes/tempDashRouter';
+// import dashboardRouter from './routes/dashboardRouter';
 
 const PORT = process.env.PORT || 3000;
 
@@ -13,9 +14,11 @@ app.use(bodyParser.json({limit: '10mb'}));
 app.use(cors());
 app.use(express.static(path.resolve(__dirname, '../client')));
 
-app.use('/results', npmRouter);
+app.use('/results', resultsRouter);
 
 app.use('/tempDash', tempDashRouter);
+
+// app.use('/dashboard', dashboardRouter);
 
 app.listen(PORT, () => {
   console.log(` 🚀 Server running on port ${PORT}`);
